backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
pgvector==0.3.2
psycopg[binary]==3.3.3
pytest==8.3.5
